+++ packages/helpers/src/clients/notion.ts
@@ -3,14 +3,19 @@
 
 import "server-only";
 import { Client } from "@notionhq/client";
-import { cachedFetch } from "../dev/cached-fetch";
+import { CachedRequestInit, cachedFetch } from "../dev/cached-fetch";
 
-export const notion = new Client({
-  auth: process.env.NOTION_API_KEY,
-  fetch: (input: RequestInfo | URL, init: RequestInit = {}) =>
-    cachedFetch(input, {
-      ...init,
-      next: { revalidate: 60 * 20, tags: ["notion-client"] },
-      devCache: { revalidate: 60 * 20, tags: ["notion-client"] },
-    }),
+export const notionWith = (init: CachedRequestInit = {}) =>
+  new Client({
+    auth: process.env.NOTION_API_KEY,
+    fetch: (input: RequestInfo | URL, fetchInit: RequestInit = {}) =>
+      cachedFetch(input, {
+        ...fetchInit,
+        ...init,
+      }),
+  });
+
+export const notion = notionWith({
+  revalidate: 60 * 20,
+  tags: ["notion-client"],
 });